Massive import refactoring
Changed JavaScript import throughout all the files.
Added a folder ./botModules/ because it became hard to distinguish between files `Bot.mjs`, `Commands.mjs`, `ButtonInteraction.mjs` and other `.mjs` "bot modules".

It appears, this improved the loading time to §play on a local machine. I will not verify much if it's true. The simple fact that it simplify code dependencies readability in the files makes it good enough to implement.