Direct use of FFMPEG to decode the video got by YTDLP
Need FFmpeg installed on the host machine, but now allow to use `seek`, useful when you want to only play part of a video

Modified the yt-dlp calls so they never try to write anything on disk, simply use it as a powerful data getter.

Removed unused `ffmpeg` and `ffmpeg-static` modules, as I directly use ffmpeg on the machine and use the cli syntax.