Use Racing Promises to fix hanging data fetching for MusicDisplayer
For the faulty or the too long to load website data extractors of yt-dlp.

Also used a Promise race for the MusicDisplayer color and gave the non-yt-dlp medias access to their Google-cached favicon to display.

Changed Play/Pause button disable behavior : will be if `currentTrack.metadata.isLive` is explicitly defined as `true`